+++ frontend_server_common/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.1.1
+- Remove dead code
+
 ## 0.1.0
 
 - Initial version
